refactor: simplify component names
