+++ includes/publisher/js/src/components/RichText/SoloRichText.jsx
@@ -1,0 +1,17 @@
+import React from "react";
+import RichTextHeader from "./RichTextHeader";
+
+const SoloRichText = ({ modelSlug, field, fieldId }) => {
+	return (
+		<>
+			<RichTextHeader modelSlug={modelSlug} field={field} />
+			<textarea
+				name={`atlas-content-modeler[${modelSlug}][${field.slug}]`}
+				id={fieldId}
+				defaultValue={field.value}
+			/>
+		</>
+	);
+};
+
+export default SoloRichText;
